Generate g2o file

+++ include/geoar/core/landmark.h
@@ -11,12 +11,12 @@ namespace geoar {
 
   class Landmark {
     public:
-      Vector3f position;
+      Vector3d position;
       cv::Mat desc;
       cv::KeyPoint kpt;
       int sightings{1};
 
-      Landmark(Vector3f &position, cv::KeyPoint &kpt, cv::Mat desc);
+      Landmark(Vector3d &position, cv::KeyPoint &kpt, cv::Mat desc);
 
       static void concatDescriptions(vector<Landmark> landmarks, cv::Mat &desc);
   };
